Move IMetricSpec/IDimensionSpec into column package
