+++ server/collector/src/main/java/org/bithon/server/collector/source/brpc/BrpcMetricCollector.java
@@ -30,14 +30,14 @@ import org.bithon.server.sink.metrics.MetricMessage;
 import org.bithon.server.sink.metrics.SchemaMetricMessage;
 import org.bithon.server.storage.datasource.DataSourceSchema;
 import org.bithon.server.storage.datasource.TimestampSpec;
-import org.bithon.server.storage.datasource.dimension.IDimensionSpec;
-import org.bithon.server.storage.datasource.dimension.StringDimensionSpec;
+import org.bithon.server.storage.datasource.column.dimension.IDimensionSpec;
+import org.bithon.server.storage.datasource.column.dimension.StringDimensionSpec;
+import org.bithon.server.storage.datasource.column.metric.IMetricSpec;
+import org.bithon.server.storage.datasource.column.metric.gauge.LongGaugeMetricSpec;
+import org.bithon.server.storage.datasource.column.metric.max.LongMaxMetricSpec;
+import org.bithon.server.storage.datasource.column.metric.min.LongMinMetricSpec;
+import org.bithon.server.storage.datasource.column.metric.sum.LongSumMetricSpec;
 import org.bithon.server.storage.datasource.input.IInputRow;
-import org.bithon.server.storage.datasource.spec.IMetricSpec;
-import org.bithon.server.storage.datasource.spec.gauge.LongGaugeMetricSpec;
-import org.bithon.server.storage.datasource.spec.max.LongMaxMetricSpec;
-import org.bithon.server.storage.datasource.spec.min.LongMinMetricSpec;
-import org.bithon.server.storage.datasource.spec.sum.LongSumMetricSpec;
 import org.springframework.util.CollectionUtils;
 
 import java.util.ArrayList;
