+++ server/storage/src/main/java/org/bithon/server/storage/StorageAutoConfiguration.java
@@ -21,10 +21,10 @@ import org.bithon.component.commons.utils.StringUtils;
 import org.bithon.server.storage.datasource.DataSourceSchema;
 import org.bithon.server.storage.datasource.DataSourceSchemaManager;
 import org.bithon.server.storage.datasource.TimestampSpec;
-import org.bithon.server.storage.datasource.dimension.IDimensionSpec;
-import org.bithon.server.storage.datasource.dimension.StringDimensionSpec;
-import org.bithon.server.storage.datasource.spec.CountMetricSpec;
-import org.bithon.server.storage.datasource.spec.sum.LongSumMetricSpec;
+import org.bithon.server.storage.datasource.column.dimension.IDimensionSpec;
+import org.bithon.server.storage.datasource.column.dimension.StringDimensionSpec;
+import org.bithon.server.storage.datasource.column.metric.CountMetricSpec;
+import org.bithon.server.storage.datasource.column.metric.sum.LongSumMetricSpec;
 import org.bithon.server.storage.event.EventStorageConfig;
 import org.bithon.server.storage.event.IEventStorage;
 import org.bithon.server.storage.meta.CacheableMetadataStorage;
