+++ server/storage/src/main/java/org/bithon/server/storage/datasource/column/metric/gauge/GaugeMetricSpec.java
@@ -14,20 +14,22 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.spec.min;
+package org.bithon.server.storage.datasource.column.metric.gauge;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
+import org.bithon.server.storage.datasource.aggregator.LongLastAggregator;
+import org.bithon.server.storage.datasource.aggregator.NumberAggregator;
+import org.bithon.server.storage.datasource.column.metric.IMetricSpec;
 import org.bithon.server.storage.datasource.query.ast.SimpleAggregateExpression;
 import org.bithon.server.storage.datasource.query.ast.SimpleAggregateExpressions;
-import org.bithon.server.storage.datasource.spec.IMetricSpec;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2021/3/16
+ * @date 2022/9/4 20:24
  */
-public abstract class MinMetricSpec implements IMetricSpec {
+public abstract class GaugeMetricSpec implements IMetricSpec {
 
     @Getter
     protected final String name;
@@ -41,16 +43,18 @@ public abstract class MinMetricSpec implements IMetricSpec {
     protected final SimpleAggregateExpression aggregateExpression;
 
     @JsonCreator
-    public MinMetricSpec(String name,
-                         String alias,
-                         String displayText) {
+    public GaugeMetricSpec(String name,
+                           String alias,
+                           String displayText) {
         this.name = name;
         this.alias = alias == null ? name : alias;
         this.displayText = displayText;
+        this.aggregateExpression = new SimpleAggregateExpressions.LastAggregateExpression(name);
+    }
 
-        // For IMetricSpec, the `name` property is the right text mapped a column in the underlying database,
-        // So the two parameters of the following ctor are all `name` properties
-        this.aggregateExpression = new SimpleAggregateExpressions.MinAggregateExpression(name);
+    @Override
+    public NumberAggregator createAggregator() {
+        return new LongLastAggregator();
     }
 
     @JsonIgnore
