+++ server/storage/src/main/java/org/bithon/server/storage/datasource/column/metric/sum/DoubleSumMetricSpec.java
@@ -14,14 +14,13 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.spec.min;
+package org.bithon.server.storage.datasource.column.metric.sum;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.bithon.server.storage.datasource.aggregator.LongMinAggregator;
+import org.bithon.server.storage.datasource.aggregator.DoubleSumAggregator;
 import org.bithon.server.storage.datasource.aggregator.NumberAggregator;
-import org.bithon.server.storage.datasource.spec.IMetricSpec;
 import org.bithon.server.storage.datasource.typing.IDataType;
 
 import javax.annotation.Nullable;
@@ -29,37 +28,37 @@ import javax.validation.constraints.NotNull;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2021/3/16
+ * @date 2020/12/23
  */
-public class LongMinMetricSpec extends MinMetricSpec {
+public class DoubleSumMetricSpec extends SumMetricSpec {
 
     @JsonCreator
-    public LongMinMetricSpec(@JsonProperty("name") @NotNull String name,
-                             @JsonProperty("alias") @Nullable String alias,
-                             @JsonProperty("displayText") @NotNull String displayText) {
+    public DoubleSumMetricSpec(@JsonProperty("name") @NotNull String name,
+                               @JsonProperty("alias") @Nullable String alias,
+                               @JsonProperty("displayText") @NotNull String displayText) {
         super(name, alias, displayText);
     }
 
     @JsonIgnore
     @Override
     public String getType() {
-        return IMetricSpec.LONG_MIN;
+        return DOUBLE_SUM;
     }
 
     @Override
     public IDataType getDataType() {
-        return IDataType.LONG;
+        return IDataType.DOUBLE;
     }
 
     @Override
     public NumberAggregator createAggregator() {
-        return new LongMinAggregator();
+        return new DoubleSumAggregator();
     }
 
     @Override
     public boolean equals(Object obj) {
-        if (obj instanceof LongMinMetricSpec) {
-            return this.name.equals(((LongMinMetricSpec) obj).name);
+        if (obj instanceof DoubleSumMetricSpec) {
+            return this.name.equals(((DoubleSumMetricSpec) obj).name);
         } else {
             return false;
         }
