+++ server/storage/src/main/java/org/bithon/server/storage/datasource/column/metric/sum/LongSumMetricSpec.java
@@ -14,12 +14,12 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.spec.gauge;
+package org.bithon.server.storage.datasource.column.metric.sum;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.bithon.server.storage.datasource.aggregator.DoubleLastAggregator;
+import org.bithon.server.storage.datasource.aggregator.LongSumAggregator;
 import org.bithon.server.storage.datasource.aggregator.NumberAggregator;
 import org.bithon.server.storage.datasource.typing.IDataType;
 
@@ -28,30 +28,39 @@ import javax.validation.constraints.NotNull;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2020/12/23
+ * @date 2020/11/30 5:38 下午
  */
-public class DoubleGaugeMetricSpec extends GaugeMetricSpec {
+public class LongSumMetricSpec extends SumMetricSpec {
 
     @JsonCreator
-    public DoubleGaugeMetricSpec(@JsonProperty("name") @NotNull String name,
-                                 @JsonProperty("alias") @Nullable String alias,
-                                 @JsonProperty("displayText") @NotNull String displayText) {
+    public LongSumMetricSpec(@JsonProperty("name") @NotNull String name,
+                             @JsonProperty("alias") @Nullable String alias,
+                             @JsonProperty("displayText") @NotNull String displayText) {
         super(name, alias, displayText);
     }
 
     @JsonIgnore
     @Override
     public String getType() {
-        return DOUBLE_LAST;
+        return LONG_SUM;
     }
 
     @Override
     public IDataType getDataType() {
-        return IDataType.DOUBLE;
+        return IDataType.LONG;
     }
 
     @Override
     public NumberAggregator createAggregator() {
-        return new DoubleLastAggregator();
+        return new LongSumAggregator();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj instanceof LongSumMetricSpec) {
+            return this.name.equals(((LongSumMetricSpec) obj).name);
+        } else {
+            return false;
+        }
     }
 }
