+++ server/storage/src/main/java/org/bithon/server/storage/datasource/column/metric/sum/SumMetricSpec.java
@@ -14,22 +14,19 @@
  *    limitations under the License.
  */
 
-package org.bithon.server.storage.datasource.spec.gauge;
+package org.bithon.server.storage.datasource.column.metric.sum;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Getter;
-import org.bithon.server.storage.datasource.aggregator.LongLastAggregator;
-import org.bithon.server.storage.datasource.aggregator.NumberAggregator;
+import org.bithon.server.storage.datasource.column.metric.IMetricSpec;
 import org.bithon.server.storage.datasource.query.ast.SimpleAggregateExpression;
 import org.bithon.server.storage.datasource.query.ast.SimpleAggregateExpressions;
-import org.bithon.server.storage.datasource.spec.IMetricSpec;
 
 /**
  * @author frank.chen021@outlook.com
- * @date 2022/9/4 20:24
+ * @date 2022/9/4 20:17
  */
-public abstract class GaugeMetricSpec implements IMetricSpec {
+public abstract class SumMetricSpec implements IMetricSpec {
 
     @Getter
     protected final String name;
@@ -37,24 +34,19 @@ public abstract class GaugeMetricSpec implements IMetricSpec {
     @Getter
     private final String alias;
 
+
     @Getter
     protected final String displayText;
 
     protected final SimpleAggregateExpression aggregateExpression;
 
-    @JsonCreator
-    public GaugeMetricSpec(String name,
-                           String alias,
-                           String displayText) {
+    public SumMetricSpec(String name,
+                         String alias,
+                         String displayText) {
         this.name = name;
         this.alias = alias == null ? name : alias;
         this.displayText = displayText;
-        this.aggregateExpression = new SimpleAggregateExpressions.LastAggregateExpression(name);
-    }
-
-    @Override
-    public NumberAggregator createAggregator() {
-        return new LongLastAggregator();
+        this.aggregateExpression = new SimpleAggregateExpressions.SumAggregateExpression(name);
     }
 
     @JsonIgnore
